[P4RT] Fix P4RT to use uint64_t to parse node-id from configDB.

+++ p4rt_app/p4runtime/p4runtime_impl.h
@@ -223,7 +223,7 @@ public:
   absl::Status RebuildSwStateAfterWarmboot(
       const std::vector<std::pair<std::string, std::string>>& port_ids,
       const std::vector<std::pair<std::string, std::string>>& cpu_queue_ids,
-      const std::optional<int>& device_id,
+      const std::optional<uint64_t>& device_id,
       const std::vector<std::string>& ports)
       ABSL_LOCKS_EXCLUDED(server_state_lock_);
